pyproject.toml
[project]
name = "tma-debat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.96.0",
    "black>=26.3.1",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.14",
    "openai>=2.32.0",
    "pg8000>=1.31.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
